bug fixes to warn system
