+++ interactions/application_command/moderation.ts
@@ -7,7 +7,11 @@ const moveCommand = new SlashCommandSubcommandBuilder()
         .setName("destination")
         .setDescription("Channel where members will be move to")
         .addChannelTypes(ChannelType.GuildVoice, ChannelType.GuildStageVoice)
-        .setRequired(true)),
+        .setRequired(true))
+    .addBooleanOption(option => option
+        .setName('everyone')
+        .setDescription('Do you want to move everyone in the VC')
+        .setRequired(false)),
 
 timeoutCommand = new SlashCommandSubcommandBuilder()
     .setName('timeout')
@@ -66,6 +70,10 @@ async function moveFunction(interaction: ChatInputCommandInteraction) {
         return;
     } else if(source.id == destination.id) {
         interaction.reply({content:`Members are already in ${destination}`, ephemeral: true})
+        return;
+    } else if(interaction.options.getBoolean('everyone')) {
+        source.members.forEach(async member => member.voice.setChannel(destination));
+        interaction.reply({content:'Members have been moved', ephemeral:true})
         return;
     }
     //interaction.deferReply({ephemeral: true});
